update: confirmの追加＆Ajax二重通信エラーの修正＆適切なアラート表示

+++ app/assets/javascripts/index.js
@@ -1,68 +1,76 @@
-$('#register-btn').click(function(){
-  console.log("sfsf");
-  $('.register').toggleClass("display_none");
-})
-
-$('#search-btn').click(function () {
-  console.log("sdsd");
-  $('.search').toggleClass("display_none");
-})
-
 $(function () {
-  $("#submit_button_create").click(function () { //送信ボタンを押すとイベントが発火します
+  function valid_data(data){
+    if ((data.title !== undefined) &&
+        (data.title[0] !== "can't be blank") &&
+        (data.title[0] !== "is too long (maximum is 100 characters)") &&
+        (data.description !== undefined) &&
+        (data.description[0] !== "can't be blank") &&
+        (data.description[0] !== "is too long (maximum is 500 characters)") &&
+        (data.price !== undefined) &&
+        (data.price[0] !== "is not a number")){
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  $("#submit_button_create").click(function () { 
     var title = $("#item_title").val();
     var description = $("#item_description").val();
     var price = $("#item_price").val();
     var image = $("#item_image").val();
     $.ajax({
-      type: 'POST', // リクエストのタイプはGETです
-      url: '/items', // URLは"/meals"を指定します
-      data: {item: { title: title, description: description, price: price, image: image } }, // コントローラへフォームの値を送信します
-      dataType: 'json' // データの型はjsonで指定します
+      type: 'POST', 
+      url: '/items', 
+      data: {item: { title: title, description: description, price: price, image: image } }, 
+      dataType: 'json' 
     })
     .done(function (data) {
-      // 通信に成功した場合の処理です
-      console.log("通信成功！！");
-      console.log(data);
-      alert("商品を登録しました！");
-      var message = `\
-                  <div class="item_row item_block" id="item_${data.id}">\
-                    <p><b>title</b>: ${data.title}</p>\
-                    <p><b>description</b>: ${data.description}</p>\
-                    <p><b>price</b>: ${data.price}</p>\
-                    <div class="center">\
-                    <a href="/items/${data.id}">Show</a>\
-                    <a href="/items/${data.id}/edit">Edit</a>\
-                    <a data-remote="true" rel="nofollow" data-method="delete" href="/items/${data.id}">Delete</a>\
+      // 通信に成功した場合の処理
+      if (valid_data(data)) {
+        alert("商品を登録しました！");
+        var message = `\
+                    <div class="item_row item_block" id="item_${data.id}">\
+                      <p><b>title</b>: ${data.title}</p>\
+                      <p><b>description</b>: ${data.description}</p>\
+                      <p><b>price</b>: ${data.price}</p>\
+                      <div class="center">\
+                      <a href="/items/${data.id}">Show</a>\
+                      <a href="/items/${data.id}/edit">Edit</a>\
+                      <a data-confirm="削除します．よろしいですか？" data-remote="true" rel="nofollow" data-method="delete" href="/items/${data.id}">Delete</a>\
+                      </div>\
                     </div>\
-                  </div>\
-                  `;
-      console.log(message);
-      $('.items').prepend(message);
-      $('.error').addClass('display_none');
+                    `;
+        // 非同期でitemを追加
+        $('.items').prepend(message);
+        // エラーメッセージは外す
+        $('.error').addClass('display_none');
+        // フォームの中身を消去する
+        $('#item_title').val("");
+        $('#item_description').val("");
+        $('#item_price').val("");
+        $('#item_image').val("");
+      } else {
+        alert("商品の登録に失敗しました。エラー内容を確認してください。");
+      }
     })
     .fail(function () {
-      // 通信に失敗した場合の処理です
-      alert('商品登録できませんでした。エラー内容を確認してください。') // alertで検索失敗の旨を表示します
+      alert('通信に失敗しました。エラー内容を確認してください。')
     })
   })
 })
 
 $(function () {
-  $("#submit_button_search").click(function () { //送信ボタンを押すとイベントが発火します
+  $("#submit_button_search").click(function () { 
     var text = $("#text").val();
     console.log(text);
     $.ajax({
-      type: 'GET', // リクエストのタイプはGETです
-      url: '/items', // URLは"/meals"を指定します
-      // data: { item: { } }, // コントローラへフォームの値を送信します
+      type: 'GET', 
+      url: '/items', 
       data: { text: text },
-      dataType: 'json' // データの型はjsonで指定します
+      dataType: 'json' 
     })
       .done(function (data) {
-        // 通信に成功した場合の処理です
-        console.log("通信成功！！！！！");
-        console.log(data);
         alert("検索結果を取得しました。");
         $('.item_row').hide();
         data.forEach(function (item) {
@@ -74,18 +82,18 @@ $(function () {
                     <div class="center">\
                     <a href="/items/${item.id}">Show</a>\
                     <a href="/items/${item.id}/edit">Edit</a>\
-                    <a data-remote="true" rel="nofollow" data-method="delete" href="/items/${item.id}">Delete</a>\
+                    <a data-confirm="削除します．よろしいですか？" data-remote="true" rel="nofollow" data-method="delete" href="/items/${item.id}">Delete</a>\
                     </div>\
                   </div>\
                   `;
           console.log(message);
           $('.items').prepend(message);
+          $("#text").val("");
         })
         $('.error').addClass('display_none');
       })
       .fail(function () {
-        // 通信に失敗した場合の処理です
-        alert('検索エラーが発生しました。') // alertで検索失敗の旨を表示します
+        alert('検索エラーが発生しました。') 
       })
   })
 })
